test(phase2): split deterministic mock order tests
